Move parser and controller tests to the right package
