+++ src/test/java/cs2103/v15_1j/jimjim/command/UndoCommandTest.java
@@ -8,7 +8,6 @@ import java.util.Stack;
 import org.junit.Before;
 import org.junit.Test;
 
-import cs2103.v15_1j.jimjim.StubStorage;
 import cs2103.v15_1j.jimjim.controller.ControllerStates;
 import cs2103.v15_1j.jimjim.model.DataLists;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
